refactor(memory): simplify recall with namespaced APIs
