+++ packages/capabilities/test/memory-lifecycle.test.ts
@@ -25,7 +25,7 @@ import {
 } from "effect";
 import { TestClock } from "effect/testing";
 
-import { MemoryAccess, recallMemory, revalidateMemoryLookup } from "../src/index.ts";
+import { MemoryAccess, Memory, revalidateMemoryLookup } from "../src/index.ts";
 
 const TestNamespace = MemoryNamespace.define({
   name: "test/memory",
@@ -79,7 +79,7 @@ const limits = MemoryRecallLimits.make({
 });
 
 const recall = (lookup = candidates) =>
-  recallMemory(
+  Memory.recall(
     [
       {
         id: "stale-cache",
@@ -120,7 +120,7 @@ describe("authoritative memory validation", () => {
             },
           }),
         );
-        const result = yield* recallMemory(
+        const result = yield* Memory.recall(
           accesses.map((bound, index) => ({
             id: `reader-${index}`,
             essential: true,
@@ -372,7 +372,7 @@ describe("authoritative memory validation", () => {
             () => Ref.update(releases, (n) => n + 1),
           ),
         );
-        const fiber = yield* recallMemory(
+        const fiber = yield* Memory.recall(
           [
             {
               id: "slow-view",
